@@ -6,52 +6,62 @@ publish_to: "none"
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.15.1 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
+  autoequal: ^0.5.1
+  bloc: ^8.1.2
+  camera: ^0.10.5+2
+  cloud_firestore: ^4.8.2
+  copy_with_extension: ^5.0.2
+  cupertino_icons: ^1.0.5
+  equatable: ^2.0.5
+  file_picker: ^5.3.2
+  firebase_analytics: ^10.4.3
+  firebase_app_check: ^0.1.5
+  firebase_auth: ^4.6.3
+  firebase_core: ^2.14.0
+  firebase_messaging: ^14.6.5
+  firebase_storage: ^11.2.4
+  firefuel: ^0.4.0+1
+  flex_color_scheme: ^7.1.2
   flutter:
     sdk: flutter
-
-  cupertino_icons: ^1.0.2
-  bloc: ^8.0.3
-  flutter_bloc: ^8.0.1
-  equatable: ^2.0.3
+  flutter_barcode_scanner: ^2.0.0
+  flutter_bloc: ^8.1.3
+  flutter_cache_manager: ^3.3.0
+  flutter_cupertino_datetime_picker: ^3.0.0
   flutter_localizations:
     sdk: flutter
-  intl: ^0.17.0
-  image_picker: ^0.8.4+5
-  flutter_cupertino_datetime_picker: ^3.0.0
-  freezed_annotation: ^1.1.0
-  path: ^1.8.0
-  path_provider: ^2.0.8
-  camera: ^0.9.4+12
-  freezed: ^1.1.1
-  flutter_cache_manager: ^3.3.0
-  jiffy: ^5.0.0
-  url_launcher: ^6.0.20
-  firebase_auth: ^3.3.9
-  firebase_core: ^1.13.1
-  cloud_firestore: ^3.1.10
-  firebase_analytics: ^9.1.2
-  flutter_login: ^3.1.0
-  shared_preferences: ^2.0.13
-  google_sign_in: ^5.2.4
-  firebase_messaging: ^11.2.10
-  go_router: ^3.0.4
-  flex_color_scheme: ^4.2.0
-  google_fonts: ^2.3.1
-  firefuel: ^0.2.1
-  sign_in_with_apple: ^3.3.0
-  crypto: ^3.0.1
-  firebase_storage: ^10.2.10
+  flutter_login: ^4.1.1
+  go_router: ^10.1.0
+  go_router_paths: ^0.2.0
+  google_sign_in: ^6.1.4
+  home_widget: ^0.3.0
+  image_picker: ^1.0.0
+  intl: ^0.18.0
+  jiffy: ^6.2.1
+  json_annotation: ^4.8.1
+  path: ^1.8.3
+  path_provider: ^2.0.15
+  shared_preferences: ^2.1.2
+  url_launcher: ^6.1.11
+  uuid: ^3.0.7
 
 dev_dependencies:
-  build_runner:
-  json_serializable: ^6.1.5
+  autoequal_gen: ^0.5.1
+  build_runner: ^2.1.7
+  copy_with_extension_gen: ^5.0.2
+  json_serializable: ^6.7.1
+
+  flutter_lints: ^2.0.2
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^1.0.0
+# For information on the generic Dart part of this file, see the
+# following page: https://dart.dev/tools/pub/pubspec
+# The following section is specific to Flutter.
+scripts: derry.yaml
 
 flutter:
   fonts:
@@ -61,3 +71,31 @@ flutter:
 
   uses-material-design: true
   generate: true
+  # To add assets to your application, add an assets section, like this:
+  # assets:
+  #   - images/a_dot_burr.jpeg
+  #   - images/a_dot_ham.jpeg
+  # An image asset can refer to one or more resolution-specific "variants", see
+  # https://flutter.dev/assets-and-images/#resolution-aware.
+  # For details regarding adding assets from package dependencies, see
+  # https://flutter.dev/assets-and-images/#from-packages
+  # To add custom fonts to your application, add a fonts section here,
+  # in this "flutter" section. Each entry in this list should have a
+  # "family" key with the font family name, and a "fonts" key with a
+  # list giving the asset and other descriptors for the font. For
+  # example:
+  # fonts:
+  #   - family: Schyler
+  #     fonts:
+  #       - asset: fonts/Schyler-Regular.ttf
+  #       - asset: fonts/Schyler-Italic.ttf
+  #         style: italic
+  #   - family: Trajan Pro
+  #     fonts:
+  #       - asset: fonts/TrajanPro.ttf
+  #       - asset: fonts/TrajanPro_Bold.ttf
+  #         weight: 700
+  #
+  # For details regarding fonts from package dependencies,
+  # see https://flutter.dev/custom-fonts/#from-packages
+  # debug token: 6843C3D7-CAA9-475F-AF65-23940B38DC69
