@@ -1,17 +1,14 @@
-import 'package:flutter/material.dart';
-import 'package:url_launcher/url_launcher.dart';
+import 'package:url_launcher/url_launcher_string.dart';
+import 'package:warranty_keeper/app_library.dart';
 
 class UrlLaunchHelper {
   UrlLaunchHelper._();
 
   static Future<bool> launchURL(String url) async {
     try {
-      if (await canLaunch(url)) {
-        await launch(
+      if (await canLaunchUrlString(url)) {
+        await launchUrlString(
           url,
-          enableJavaScript: true,
-          forceSafariVC: true,
-          forceWebView: true,
         );
         return true;
       }
