@@ -8,10 +8,11 @@ class WarrantyTextField extends StatefulWidget {
   final int? maxLength;
   final int? currentLength;
   final MaxLengthEnforcement? maxLengthEnforcement;
-  final TextInputType textInputType;
+  final TextInputType? textInputType;
   final String hintText;
   final String initialText;
   final String initialValue;
+  final String? errorText;
   final bool isRequired;
   final bool hasAutocorrect;
   final bool isTextObscured;
@@ -23,12 +24,15 @@ class WarrantyTextField extends StatefulWidget {
   final DateTime? initialDateTime;
   final DateTime? startDateTime;
   final DateTime? endDateTime;
+  final List<TextInputFormatter>? inputFormatter;
   final Function(String)? onChanged;
   final VoidCallback? onObscuredTap;
   const WarrantyTextField({
     Key? key,
     required this.maxLength,
     this.currentLength,
+    this.errorText,
+    this.inputFormatter,
     required this.maxLengthEnforcement,
     required this.textInputType,
     required this.hintText,
@@ -65,9 +69,11 @@ class WarrantyTextField extends StatefulWidget {
     this.maxLength,
     this.maxLengthEnforcement,
     this.currentLength,
+    this.errorText,
   })  : initialText = initialDateTime != null
             ? DateFormat('MM/dd/yyyy').format(initialDateTime)
             : '',
+        inputFormatter = [],
         textInputType = TextInputType.none,
         isDate = true,
         maxLines = 1,
@@ -77,13 +83,47 @@ class WarrantyTextField extends StatefulWidget {
         hasAutocorrect = false,
         super(key: key);
 
+  WarrantyTextField.dob({
+    this.initialDateTime,
+    required this.initialValue,
+    required this.hintText,
+    required this.onChanged,
+    required this.isRequired,
+    this.onTap,
+    key,
+    this.maxLength,
+    this.maxLengthEnforcement,
+    this.currentLength,
+    this.errorText,
+  })  : initialText = initialDateTime != null
+            ? DateFormat('MM/dd/yyyy').format(initialDateTime)
+            : '',
+        textInputType = TextInputType.none,
+        isDate = true,
+        inputFormatter = [],
+        maxLines = 1,
+        isTextObscured = false,
+        onObscuredTap = null,
+        isObscuredFunction = false,
+        hasAutocorrect = false,
+        isLifeTime = false,
+        startDateTime = DateTime(1920),
+        endDateTime = DateTime.now().subtract(
+          const Duration(
+            days: 4748,
+          ),
+        ),
+        super(key: key);
+
   const WarrantyTextField.general({
     this.maxLength,
+    this.errorText,
     this.maxLengthEnforcement,
     required this.isRequired,
     required this.initialValue,
     this.onTap,
     required this.hintText,
+    this.inputFormatter,
     required this.onChanged,
     this.initialDateTime,
     this.startDateTime,
@@ -101,13 +141,41 @@ class WarrantyTextField extends StatefulWidget {
         textInputType = TextInputType.text,
         super(key: key);
 
+  const WarrantyTextField.number({
+    this.maxLength,
+    this.errorText,
+    this.maxLengthEnforcement,
+    required this.isRequired,
+    required this.initialValue,
+    this.onTap,
+    required this.hintText,
+    required this.onChanged,
+    this.initialDateTime,
+    this.startDateTime,
+    this.endDateTime,
+    this.inputFormatter,
+    key,
+    this.currentLength,
+  })  : initialText = '',
+        isDate = false,
+        isTextObscured = false,
+        onObscuredTap = null,
+        isObscuredFunction = false,
+        isLifeTime = false,
+        maxLines = 1,
+        hasAutocorrect = true,
+        textInputType = TextInputType.number,
+        super(key: key);
+
   const WarrantyTextField.email({
     this.maxLength,
+    this.errorText,
     this.maxLengthEnforcement,
     required this.isRequired,
     required this.initialValue,
     this.onTap,
     required this.onChanged,
+    this.inputFormatter,
     this.initialDateTime,
     this.startDateTime,
     this.endDateTime,
@@ -127,9 +195,11 @@ class WarrantyTextField extends StatefulWidget {
 
   const WarrantyTextField.obscured({
     this.maxLength,
+    this.errorText,
     this.maxLengthEnforcement,
     required this.isRequired,
     required this.initialValue,
+    this.inputFormatter,
     this.onTap,
     required this.hintText,
     required this.onChanged,
@@ -153,9 +223,11 @@ class WarrantyTextField extends StatefulWidget {
     required this.isRequired,
     required this.initialValue,
     this.onTap,
+    this.errorText,
     required this.hintText,
     required this.onChanged,
     this.initialDateTime,
+    this.inputFormatter,
     this.startDateTime,
     this.endDateTime,
     key,
@@ -175,7 +247,9 @@ class WarrantyTextField extends StatefulWidget {
 
   const WarrantyTextField.form({
     this.onTap,
+    this.errorText,
     required this.hintText,
+    this.inputFormatter,
     required this.initialValue,
     required this.onChanged,
     this.initialDateTime,
@@ -215,7 +289,7 @@ class _WarrantyTextFieldState extends State<WarrantyTextField> {
   @override
   Widget build(BuildContext context) {
     //Used for Date Picker
-    String _dateToString(DateTime date) {
+    String dateToString(DateTime date) {
       final DateFormat formatter = DateFormat('MM/dd/yyyy');
       return formatter.format(date);
     }
@@ -231,7 +305,7 @@ class _WarrantyTextFieldState extends State<WarrantyTextField> {
               mode: CupertinoDatePickerMode.date,
               onDateTimeChanged: (datePicked) {
                 if (datePicked != widget.initialDateTime) {
-                  controller.text = _dateToString(datePicked);
+                  controller.text = dateToString(datePicked);
                   widget.onChanged?.call(controller.text);
                 }
               },
@@ -248,19 +322,20 @@ class _WarrantyTextFieldState extends State<WarrantyTextField> {
 
     Future<void> buildMaterialDatePicker() async {
       final datePicked = await showDatePicker(
+        keyboardType: TextInputType.text,
         context: context,
         initialDate: widget.initialDateTime ?? DateTime.now(),
         firstDate: widget.startDateTime ?? DateTime(2000),
         lastDate: widget.endDateTime ?? DateTime(2050),
-        initialEntryMode: DatePickerEntryMode.input,
+        initialEntryMode: DatePickerEntryMode.calendar,
       );
       if (datePicked != null && datePicked != widget.initialDateTime) {
-        controller.text = _dateToString(datePicked);
+        controller.text = dateToString(datePicked);
         widget.onChanged?.call(controller.text);
       }
     }
 
-    void _selectDate() {
+    void selectDate() {
       switch (context.themeData.platform) {
         case TargetPlatform.iOS:
           buildCupertinoDatePicker();
@@ -285,16 +360,15 @@ class _WarrantyTextFieldState extends State<WarrantyTextField> {
             maxLengthEnforcement: widget.maxLengthEnforcement,
             keyboardType: widget.textInputType,
             enabled: widget.isLifeTime ? false : true,
-            inputFormatters: [
-              if (widget.isDate) DateTextFormatter(),
-            ],
+            inputFormatters: widget.inputFormatter,
             controller: controller,
             onChanged: widget.onChanged,
             onTap: () {
-              widget.isDate ? _selectDate() : widget.onTap;
+              widget.isDate ? selectDate() : widget.onTap;
             },
             maxLines: widget.maxLines ?? 1,
             decoration: InputDecoration(
+              errorText: widget.errorText,
               suffixIcon: widget.isTextObscured
                   ? GestureDetector(
                       onTap: widget.onObscuredTap,
