@@ -1,5 +1,5 @@
 // File generated by FlutterFire CLI.
-// ignore_for_file: lines_longer_than_80_chars
+// ignore_for_file: lines_longer_than_80_chars, avoid_classes_with_only_static_members
 import 'package:firebase_core/firebase_core.dart' show FirebaseOptions;
 import 'package:flutter/foundation.dart'
     show defaultTargetPlatform, kIsWeb, TargetPlatform;
@@ -17,12 +17,8 @@ import 'package:flutter/foundation.dart'
 class DefaultFirebaseOptions {
   static FirebaseOptions get currentPlatform {
     if (kIsWeb) {
-      throw UnsupportedError(
-        'DefaultFirebaseOptions have not been configured for web - '
-        'you can reconfigure this by running the FlutterFire CLI again.',
-      );
+      return web;
     }
-    // ignore: missing_enum_constant_in_switch
     switch (defaultTargetPlatform) {
       case TargetPlatform.android:
         return android;
@@ -33,31 +29,52 @@ class DefaultFirebaseOptions {
           'DefaultFirebaseOptions have not been configured for macos - '
           'you can reconfigure this by running the FlutterFire CLI again.',
         );
+      case TargetPlatform.windows:
+        throw UnsupportedError(
+          'DefaultFirebaseOptions have not been configured for windows - '
+          'you can reconfigure this by running the FlutterFire CLI again.',
+        );
+      case TargetPlatform.linux:
+        throw UnsupportedError(
+          'DefaultFirebaseOptions have not been configured for linux - '
+          'you can reconfigure this by running the FlutterFire CLI again.',
+        );
+      default:
+        throw UnsupportedError(
+          'DefaultFirebaseOptions are not supported for this platform.',
+        );
     }
-
-    throw UnsupportedError(
-      'DefaultFirebaseOptions are not supported for this platform.',
-    );
   }
+
+  static const FirebaseOptions web = FirebaseOptions(
+    apiKey: 'AIzaSyC0Vp9okbDB_CVPsef7U0CtiPrPzUN_ZTs',
+    appId: '1:335544571045:web:09e6de694e11e3e20e694c',
+    messagingSenderId: '335544571045',
+    projectId: 'warranty-tracker-a6492',
+    authDomain: 'warranty-tracker-a6492.firebaseapp.com',
+    databaseURL: 'https://warranty-tracker-a6492-default-rtdb.firebaseio.com',
+    storageBucket: 'warranty-tracker-a6492.appspot.com',
+    measurementId: 'G-C0CX1EHLWM',
+  );
 
   static const FirebaseOptions android = FirebaseOptions(
     apiKey: 'AIzaSyA8P5RiHEZn7HY6co73iZwB7xo94VrBJVU',
     appId: '1:335544571045:android:cc3308192b89c8770e694c',
     messagingSenderId: '335544571045',
     projectId: 'warranty-tracker-a6492',
+    databaseURL: 'https://warranty-tracker-a6492-default-rtdb.firebaseio.com',
     storageBucket: 'warranty-tracker-a6492.appspot.com',
   );
 
   static const FirebaseOptions ios = FirebaseOptions(
     apiKey: 'AIzaSyAGrVBTY0phaqM9Ld4oq2Pq1HAr7DNPSUc',
-    appId: '1:335544571045:ios:7130e5593927804d0e694c',
+    appId: '1:335544571045:ios:e4acd72e29d7130d0e694c',
     messagingSenderId: '335544571045',
     projectId: 'warranty-tracker-a6492',
+    databaseURL: 'https://warranty-tracker-a6492-default-rtdb.firebaseio.com',
     storageBucket: 'warranty-tracker-a6492.appspot.com',
-    androidClientId:
-        '335544571045-btg9lfgm4vpsid4lhmq4sebpjjbvcmop.apps.googleusercontent.com',
-    iosClientId:
-        '335544571045-qij4c9hhgqd73gfa7g3ftic0so394kj7.apps.googleusercontent.com',
-    iosBundleId: 'com.warranty_tracker.underDevelopment.app',
+    androidClientId: '335544571045-btg9lfgm4vpsid4lhmq4sebpjjbvcmop.apps.googleusercontent.com',
+    iosClientId: '335544571045-1sms6mc3ha35inr00vsr3gtj0563v5qq.apps.googleusercontent.com',
+    iosBundleId: 'com.warrantyKeeper.underDevelopment',
   );
 }
