@@ -1,29 +1,29 @@
-import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_app_check/firebase_app_check.dart';
 import 'package:firebase_core/firebase_core.dart';
-import 'package:firefuel/firefuel.dart';
 import 'package:flex_color_scheme/flex_color_scheme.dart';
-import 'package:google_fonts/google_fonts.dart';
+import 'package:flutter_localizations/flutter_localizations.dart';
+import 'package:warranty_keeper/data/data_sources.dart/firebase_data_source.dart';
 import 'package:warranty_keeper/data/repositories/auth_repository.dart';
 import 'package:warranty_keeper/modules/cubit/auth/auth_cubit.dart';
-import 'package:warranty_keeper/modules/cubit/current_warranties/current_warranties_cubit.dart';
-import 'package:warranty_keeper/modules/cubit/home/home_cubit.dart';
-import 'package:warranty_keeper/modules/cubit/main/main_cubit.dart';
-import 'package:warranty_keeper/modules/cubit/nav_cubit/nav_cubit.dart';
-import 'package:warranty_keeper/modules/cubit/new_warranty/new_warranty_cubit.dart';
-import 'package:warranty_keeper/modules/cubit/settings/settings_cubit.dart';
+import 'package:warranty_keeper/modules/cubit/warranties/warranties_cubit.dart';
 import 'package:warranty_keeper/modules/cubit/warranty_details/warranty_details_cubit.dart';
-
-import 'package:warranty_keeper/routes/app_routes.dart';
-
-import 'app_library.dart';
+import 'data/repositories/warranty_repository.dart';
 import 'firebase_options.dart';
 
+import 'package:warranty_keeper/routes/go_routes.dart';
+
+import 'app_library.dart';
+
 void main() async {
+  // const webRecaptchaSiteKey = '';
   WidgetsFlutterBinding.ensureInitialized();
   await Firebase.initializeApp(
     options: DefaultFirebaseOptions.currentPlatform,
   );
-  Firefuel.initialize(FirebaseFirestore.instance);
+  await FirebaseAPI().initNotifications();
+  await FirebaseAppCheck.instance.activate(
+    androidProvider: AndroidProvider.debug,
+  );
   runApp(const MyApp());
 }
 
@@ -34,133 +34,53 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    // final streamController = StreamController();
-    final _router = appRoutes(/* streamController */);
+    ThemeMode themeMode = ThemeMode.system;
+
+    // const usedScheme = FlexScheme.deepBlue;
+    // const usedScheme = FlexScheme.flutterDash;
+    const usedScheme = FlexScheme.blue;
+    // const usedScheme = FlexScheme.aquaBlue;
+    // const usedScheme = FlexScheme.brandBlue;
+    // const usedScheme = FlexScheme.hippieBlue;
+    // const usedScheme = FlexScheme.bahamaBlue;
 
     return MultiBlocProvider(
       providers: [
-        BlocProvider<MainCubit>(
-          create: (context) => MainCubit(),
+        BlocProvider<AuthCubit>(
+          create: (context) => AuthCubit(FirebaseAuthRepository()),
         ),
-        BlocProvider<CurrentWarrantiesCubit>(
-          create: (context) => CurrentWarrantiesCubit(),
-        ),
-        BlocProvider<HomeCubit>(
-          create: (context) => HomeCubit(),
-        ),
-        BlocProvider<NavCubit>(
-          create: (context) => NavCubit(),
-        ),
-        BlocProvider<NewWarrantyCubit>(
-          create: (context) => NewWarrantyCubit(),
+        BlocProvider<WarrantiesCubit>(
+          create: (context) => WarrantiesCubit(
+            dataRepository: DataRepository(),
+            authRepository: FirebaseAuthRepository(),
+          ),
         ),
         BlocProvider<WarrantyDetailsCubit>(
           create: (context) => WarrantyDetailsCubit(),
         ),
-        BlocProvider<SettingsCubit>(
-          create: (context) => SettingsCubit(),
-        ),
-        BlocProvider<AuthCubit>(
-          create: (context) => AuthCubit(FirebaseAuthRepository()),
-        ),
       ],
       child: MaterialApp.router(
-        routerDelegate: _router.routerDelegate,
-        routeInformationParser: _router.routeInformationParser,
-        localizationsDelegates: AppLocalizations.localizationsDelegates,
-        supportedLocales: AppLocalizations.supportedLocales,
-        title: 'Warranty Tracker or Warranty Notifier',
-        darkTheme: FlexThemeData.dark(
-          colors: const FlexSchemeColor(
-            primary: Color(0xff5b7cb5),
-            primaryVariant: Color(0xff5a6f94),
-            secondary: Color(0xffc15c5c),
-            secondaryVariant: Color(0xffa15a5a),
-            appBarColor: Color(0xffa15a5a),
-            error: null,
-          ),
-          surfaceMode: FlexSurfaceMode.highScaffoldLowSurface,
-          blendLevel: 18,
-          appBarStyle: FlexAppBarStyle.background,
-          appBarOpacity: 0.95,
-          appBarElevation: 0,
-          transparentStatusBar: true,
-          tabBarStyle: FlexTabBarStyle.forAppBar,
-          tooltipsMatchBackground: true,
-          swapColors: false,
-          darkIsTrueBlack: false,
-          useSubThemes: true,
-          visualDensity: FlexColorScheme.comfortablePlatformDensity,
-          fontFamily: GoogleFonts.notoSans().fontFamily,
-          subThemesData: const FlexSubThemesData(
-            useTextTheme: true,
-            fabUseShape: true,
-            interactionEffects: true,
-            bottomNavigationBarElevation: 0,
-            bottomNavigationBarOpacity: 0.95,
-            navigationBarOpacity: 0.95,
-            navigationBarMutedUnselectedText: true,
-            navigationBarMutedUnselectedIcon: true,
-            inputDecoratorIsFilled: true,
-            inputDecoratorBorderType: FlexInputBorderType.outline,
-            inputDecoratorUnfocusedHasBorder: true,
-            blendOnColors: true,
-            blendTextTheme: true,
-            popupMenuOpacity: 0.95,
-          ),
-        ),
+        localizationsDelegates: const [
+          AppLocalizations.delegate,
+          GlobalMaterialLocalizations.delegate,
+          GlobalWidgetsLocalizations.delegate,
+          GlobalCupertinoLocalizations.delegate,
+        ],
+        supportedLocales: const [
+          Locale('en'), // English
+          Locale('es'), // Spanish
+        ],
+        title: 'Warranty Tracker',
         theme: FlexThemeData.light(
-          colors: const FlexSchemeColor(
-            primary: Color(0xff04368e),
-            primaryVariant: Color(0xff03235c),
-            secondary: Color(0xffa00505),
-            secondaryVariant: Color(0xff6f0303),
-            appBarColor: Color(0xff6f0303),
-            error: null,
-          ),
-          surfaceMode: FlexSurfaceMode.highScaffoldLowSurface,
-          blendLevel: 18,
-          appBarStyle: FlexAppBarStyle.primary,
-          appBarOpacity: 0.95,
-          appBarElevation: 0,
-          transparentStatusBar: true,
-          tabBarStyle: FlexTabBarStyle.forAppBar,
-          tooltipsMatchBackground: true,
-          swapColors: false,
-          lightIsWhite: false,
-          useSubThemes: true,
-          visualDensity: FlexColorScheme.comfortablePlatformDensity,
-          fontFamily: GoogleFonts.notoSans().fontFamily,
-          subThemesData: const FlexSubThemesData(
-            useTextTheme: true,
-            fabUseShape: true,
-            interactionEffects: true,
-            bottomNavigationBarElevation: 0,
-            bottomNavigationBarOpacity: 0.95,
-            navigationBarOpacity: 0.95,
-            navigationBarMutedUnselectedText: true,
-            navigationBarMutedUnselectedIcon: true,
-            inputDecoratorIsFilled: true,
-            inputDecoratorBorderType: FlexInputBorderType.outline,
-            inputDecoratorUnfocusedHasBorder: true,
-            blendOnColors: true,
-            blendTextTheme: true,
-            popupMenuOpacity: 0.95,
-          ),
+          scheme: usedScheme,
+          appBarElevation: 0.5,
         ),
-        /*  FlexColorScheme.light(scheme: FlexScheme.amber).toTheme, */
-        /* ThemeData(
-          primarySwatch: Colors.blue,
-        ).copyWith(
-          pageTransitionsTheme: const PageTransitionsTheme(
-            builders: <TargetPlatform, PageTransitionsBuilder>{
-              TargetPlatform.android: CupertinoPageTransitionsBuilder(),
-              TargetPlatform.iOS: CupertinoPageTransitionsBuilder(),
-            },
-          ),
-        ), */
-        // home: const LoginView(),
-        // routes: appRoutes,
+        darkTheme: FlexThemeData.dark(
+          scheme: usedScheme,
+          appBarElevation: 2,
+        ),
+        themeMode: themeMode,
+        routerConfig: goRoutes,
       ),
     );
   }
