@@ -1,28 +1,38 @@
+import 'package:autoequal/autoequal.dart';
 import 'package:bloc/bloc.dart';
-import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:copy_with_extension/copy_with_extension.dart';
+import 'package:equatable/equatable.dart';
 
 part 'login_state.dart';
-part 'login_cubit.freezed.dart';
+part 'login_cubit.g.dart';
 
 class LoginCubit extends Cubit<LoginState> {
-  LoginCubit() : super(const LoginState.initial());
+  LoginCubit() : super(const _Initial()) {
+    setInitial();
+  }
+
+  setInitial() {
+    emit(
+      const _LoggingIn(
+        email: '',
+        password: '',
+        isObscured: true,
+      ),
+    );
+  }
 
   toggleObscurity() {
     emit(
-      state.copyWith(
-        isObscured: !state.isObscured,
-        email: state.email,
-        password: state.password,
+      state.asLoggingIn.copyWith(
+        isObscured: !state.asLoggingIn.isObscured,
       ),
     );
   }
 
   changeEmail(String email) {
     emit(
-      state.copyWith(
+      state.asLoggingIn.copyWith(
         email: email,
-        isObscured: state.isObscured,
-        password: state.password,
       ),
     );
     enabledLogin();
@@ -30,9 +40,7 @@ class LoginCubit extends Cubit<LoginState> {
 
   changePassword(String password) {
     emit(
-      state.copyWith(
-        email: state.email,
-        isObscured: state.isObscured,
+      state.asLoggingIn.copyWith(
         password: password,
       ),
     );
@@ -40,8 +48,8 @@ class LoginCubit extends Cubit<LoginState> {
   }
 
   bool enabledLogin() {
-    if (state.email.isEmpty) return false;
-    if (state.password.isEmpty) return false;
+    if (state.asLoggingIn.email.isEmpty) return false;
+    if (state.asLoggingIn.password.isEmpty) return false;
     return true;
   }
 }
