@@ -1,14 +1,10 @@
-import 'package:go_router/go_router.dart';
 import 'package:warranty_keeper/app_library.dart';
-import 'package:warranty_keeper/modules/cubit/current_warranties/current_warranties_cubit.dart';
-import 'package:warranty_keeper/modules/cubit/new_warranty/new_warranty_cubit.dart';
+import 'package:warranty_keeper/modules/cubit/warranties/warranties_cubit.dart';
 import 'package:warranty_keeper/modules/cubit/warranty_details/warranty_details_cubit.dart';
 import 'package:warranty_keeper/presentation/current_warranties/widgets/current_widget_card.dart';
-import 'package:warranty_keeper/routes/paths.dart';
+import 'package:warranty_keeper/presentation/loading/widgets/triangle_loading_indicator.dart';
 
 class CurrentWarrantiesView extends StatelessWidget {
-  static const routeName = '/currentWarrantiesView';
-
   const CurrentWarrantiesView({Key? key}) : super(key: key);
 
   @override
@@ -30,33 +26,42 @@ class _Content extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final currentCubit = context.read<CurrentWarrantiesCubit>();
-    final newRead = context.read<NewWarrantyCubit>();
-    final detailsCubit = context.read<WarrantyDetailsCubit>();
     return SafeArea(
       child: Padding(
         padding: const EdgeInsets.symmetric(horizontal: 25.0, vertical: 15),
-        child: BlocBuilder<CurrentWarrantiesCubit, CurrentWarrantiesState>(
+        child: BlocBuilder<WarrantiesCubit, WarrantiesState>(
           builder: (context, state) {
-            return (state.warrantyInfoList.isEmpty)
+            if (state.isError) {
+              return const Text('Unable to find warranties');
+            } else if (state.isLoading) {
+              return const TriangleLoadingIndicator();
+            }
+            final list =
+                context.watch<WarrantiesCubit>().state.asReady.warrantyInfo;
+            return (list.isEmpty)
                 ? Text(context.appLocalizations.noCurrentWarranties)
                 : ListView.builder(
-                    itemCount: state.warrantyInfoList.length,
+                    itemCount: list.length,
                     itemBuilder: (context, index) {
                       return CurrentWidgetCard(
                         onSelect: () {
-                          context.push(Paths.warranty.selected.path);
-                          detailsCubit.selectedWarrantyInitial(
-                            state.warrantyInfoList[index],
-                          );
+                          context
+                              .read<WarrantyDetailsCubit>()
+                              .selectedWarrantyInitial(
+                                list[index],
+                              );
+                          context.push(Paths.home.warranties.details.path);
                         },
                         onEdit: () {
-                          newRead.editWarrantyInitial(
-                              state.warrantyInfoList[index]);
-                          context.push(Paths.warranty.newWarranty.path);
+                          // context
+                          //     .read<WarrantyCubit>()
+                          //     .editWarrantyInitial(list[index]);
+                          // context.pushNamed(Paths.home.newWarranty.path);
                         },
-                        onRemove: () => currentCubit.removeWarranty(index),
-                        warrantyInfo: state.warrantyInfoList[index],
+                        onRemove: () => context
+                            .read<WarrantiesCubit>()
+                            .removeWarranty(index),
+                        warrantyInfo: list[index],
                       );
                     },
                   );
