@@ -1,26 +1,16 @@
-import 'package:go_router/go_router.dart';
-
 import 'package:warranty_keeper/app_library.dart';
 import 'package:warranty_keeper/modules/cubit/auth/auth_cubit.dart';
 import 'package:warranty_keeper/modules/cubit/login/login_cubit.dart';
-import 'package:warranty_keeper/routes/paths.dart';
+import 'package:warranty_keeper/presentation/login/widgets/get_screen_height.dart';
 import 'package:warranty_keeper/widgets/sign_in_options_icons.dart';
-import 'package:warranty_keeper/widgets/warranty_button.dart';
-import 'package:warranty_keeper/widgets/warranty_textfield.dart';
-
-import 'widgets/login_options.dart';
 
 class LoginView extends StatelessWidget {
   const LoginView({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      resizeToAvoidBottomInset: false,
-      body: BlocProvider(
-        create: (context) => LoginCubit(),
-        child: const _Content(),
-      ),
+    return const Scaffold(
+      body: _Content(),
     );
   }
 }
@@ -32,69 +22,61 @@ class _Content extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final loginCubit = context.read<LoginCubit>();
-    final authCubit = context.read<AuthCubit>();
     final appLocalizations = context.appLocalizations;
 
-    errorSnackBar({required String message}) {
-      return SnackBar(
-        content: Text(message),
-      );
-    }
+    final authCubit = context.watch<AuthCubit>();
 
-    return SafeArea(
-      child: Padding(
-        padding: const EdgeInsets.symmetric(vertical: 8.0, horizontal: 25),
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.spaceAround,
-          children: [
-            Container(
-              decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(10),
-                color: context.colorScheme.primary,
-              ),
-              padding: const EdgeInsets.all(8),
-              child: Text(
-                appLocalizations.mainTitle,
-                style: context.textTheme.headline4
-                    ?.copyWith(color: context.colorScheme.onPrimary),
-              ),
-            ),
-            Flexible(
-              child: ScrollConfiguration(
-                behavior: const ScrollBehavior().copyWith(
-                  overscroll: false,
-                ),
-                child: SingleChildScrollView(
-                  physics: const ClampingScrollPhysics(),
-                  child: Column(
+    return BlocProvider<LoginCubit>(
+      create: (context) => LoginCubit(),
+      child: BlocListener<AuthCubit, AuthState>(
+        listener: (context, state) {
+          if (state.isAuthenticated) {
+            GoRouter.of(context).refresh();
+          }
+        },
+        child: SafeArea(
+          minimum: const EdgeInsets.symmetric(horizontal: 25),
+          child: SingleChildScrollView(
+            physics: const ClampingScrollPhysics(),
+            child: SizedBox(
+              height: getAvailableScreenHeight(context),
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.spaceAround,
+                children: [
+                  Container(
+                    decoration: BoxDecoration(
+                      borderRadius: BorderRadius.circular(10),
+                      color: context.colorScheme.primary,
+                    ),
+                    padding: const EdgeInsets.all(8),
+                    child: Text(
+                      appLocalizations.mainTitle,
+                      style: context.textTheme.headlineMedium?.copyWith(
+                        color: context.colorScheme.onPrimary,
+                      ),
+                    ),
+                  ),
+                  Column(
                     children: [
-                      BlocListener<AuthCubit, AuthState>(
-                        listener: (context, state) {
-                          // state.map(
-                          //   initial: initial,
-                          //   loading: loading,
-                          //   authenticated: authenticated,
-                          //   notAuthenticated: notAuthenticated,
-                          //   error: error,
-                          //   passwordRequestSubmitted: passwordRequestSubmitted,
-                          //   firstRun: firstRun,
-                          //   personalDataUpdated: personalDataUpdated,
-                          // );
-                          state.mapOrNull(
-                            error: (error) {
-                              ScaffoldMessenger.of(context).showSnackBar(
-                                errorSnackBar(message: error.message),
-                              );
+                      if (authCubit.state.isError)
+                        RichText(
+                          text: TextSpan(
+                            text: authCubit.state.asError.message,
+                            style: TextStyle(
+                              color: context.colorScheme.error,
+                            ),
+                          ),
+                        ),
+                      BlocBuilder<LoginCubit, LoginState>(
+                        builder: (context, state) {
+                          return WarrantyTextField.email(
+                            isRequired: false,
+                            initialValue: '',
+                            onChanged: (email) {
+                              context.read<LoginCubit>().changeEmail(email);
                             },
                           );
                         },
-                        child: const SizedBox(),
-                      ),
-                      WarrantyTextField.email(
-                        isRequired: false,
-                        initialValue: '',
-                        onChanged: loginCubit.changeEmail,
                       ),
                       Column(
                         crossAxisAlignment: CrossAxisAlignment.end,
@@ -102,83 +84,90 @@ class _Content extends StatelessWidget {
                           BlocBuilder<LoginCubit, LoginState>(
                             builder: (context, state) {
                               return WarrantyTextField.obscured(
-                                onObscuredTap: loginCubit.toggleObscurity,
-                                isObscuredFunction: state.isObscured,
+                                onObscuredTap: () => context
+                                    .read<LoginCubit>()
+                                    .toggleObscurity(),
+                                isObscuredFunction:
+                                    state.asLoggingIn.isObscured,
                                 isRequired: false,
                                 initialValue: '',
                                 hintText: 'Password',
-                                onChanged: loginCubit.changePassword,
+                                onChanged: (password) {
+                                  context
+                                      .read<LoginCubit>()
+                                      .changePassword(password);
+                                },
                               );
                             },
                           ),
                           TextButton(
                             child: const Text('Forgot Password?'),
-                            onPressed: () => context.push(Paths.warranty.name),
+                            onPressed: () {
+                              context.push(Paths.login.forgotPassword.path);
+                            },
                           ),
                         ],
                       ),
                     ],
                   ),
-                ),
-              ),
-            ),
-            Column(
-              children: [
-                BlocConsumer<AuthCubit, AuthState>(
-                  listener: (context, state) {
-                    state.mapOrNull(
-                      authenticated: (value) => context.go(Paths.warranty.name),
-                    );
-                  },
-                  builder: (context, state) {
-                    return BlocBuilder<LoginCubit, LoginState>(
-                      builder: (context, loginState) {
-                        return WarrantyElevatedButton.loading(
-                          isLoading: state ==
-                              const AuthState.loading(
-                                  loadingState: LoadingState.email),
-                          onPressed: () async {
-                            await authCubit.login(
-                              loginCubit.state.email,
-                              loginCubit.state.password,
-                            );
-                          },
-                          text: 'Login',
-                          isEnabled: loginCubit.enabledLogin(),
-                        );
-                      },
-                    );
-                  },
-                ),
-                TextButton(
-                  child: const Text('Sign up'),
-                  onPressed: () => context.push(Paths.login.signup.path),
-                ),
-                BlocBuilder<AuthCubit, AuthState>(
-                  builder: (context, state) {
-                    return Platform.isIOS
-                        ? LoginOptions(
-                            state: state,
-                            widget: WarrantyElevatedButton.iconLoading(
-                              onPressed: () async {
-                                //TODO: ADD AFTER APPLE ID IS SET UP
-                                // await authCubit.loginWithApple();
-                              },
-                              isLoading: state ==
-                                  const AuthState.loading(
-                                      loadingState: LoadingState.apple),
-                              isEnabled: true,
+                  Column(
+                    children: [
+                      BlocBuilder<LoginCubit, LoginState>(
+                        builder: (context, state) {
+                          return WarrantyElevatedButton(
+                            isLoading: authCubit.state.isLoading &&
+                                authCubit.state.asLoading.loginType ==
+                                    LoginType.email,
+                            onPressed: () => context.read<AuthCubit>().login(
+                                  email: state.asLoggingIn.email,
+                                  password: state.asLoggingIn.password,
+                                ),
+                            text: 'Login',
+                            isEnabled: !authCubit.state.isLoading,
+                          );
+                        },
+                      ),
+                      TextButton(
+                        child: const Text('Sign up'),
+                        onPressed: () {
+                          context.read<AuthCubit>().setInitial();
+                          context.push(
+                            Paths.login.register.path,
+                          );
+                        },
+                      ),
+                      Row(
+                        mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                        children: [
+                          WarrantyElevatedButton.iconLoading(
+                            onPressed: () =>
+                                context.read<AuthCubit>().loginWithGoogle(),
+                            isLoading: authCubit.state.isLoading &&
+                                authCubit.state.asLoading.loginType ==
+                                    LoginType.google,
+                            isEnabled: !authCubit.state.isLoading,
+                            widget: const Icon(SignInOptions.google),
+                          ),
+                          Padding(
+                            padding:
+                                const EdgeInsets.symmetric(horizontal: 3.0),
+                            child: WarrantyElevatedButton.iconLoading(
+                              onPressed: () {},
+                              isLoading: authCubit.state.isLoading &&
+                                  authCubit.state.asLoading.loginType ==
+                                      LoginType.apple,
+                              isEnabled: !authCubit.state.isLoading,
                               widget: const Icon(SignInOptions.apple),
                             ),
-                          )
-                        : LoginOptions(
-                            state: state,
-                          );
-                  },
-                ),
-              ],
-            )
-          ],
+                          ),
+                        ],
+                      )
+                    ],
+                  )
+                ],
+              ),
+            ),
+          ),
         ),
       ),
     );
