@@ -3,65 +3,218 @@
 part of 'warranty_info.dart';
 
 // **************************************************************************
-// JsonSerializableGenerator
+// AutoequalGenerator
 // **************************************************************************
 
-WarrantyInfo _$WarrantyInfoFromJson(Map json) => WarrantyInfo(
-      warrantyId: json['warrantyId'] as String,
-      warrantyState:
-          $enumDecodeNullable(_$WarrantyStateEnumMap, json['warrantyState']) ??
-              WarrantyState.initial,
-      name: json['name'] as String?,
-      purchaseDate: json['purchaseDate'] == null
-          ? null
-          : DateTime.parse(json['purchaseDate'] as String),
-      warrantyWebsite: json['warrantyWebsite'] as String?,
-      endOfWarranty: json['endOfWarranty'] == null
-          ? null
-          : DateTime.parse(json['endOfWarranty'] as String),
-      reminderDate: json['reminderDate'] == null
-          ? null
-          : DateTime.parse(json['reminderDate'] as String),
-      details: json['details'] as String?,
-      image: json['image'] as String?,
-      imageUrl: json['imageUrl'] as String?,
-      receiptImage: json['receiptImage'] as String?,
-      receiptImageUrl: json['receiptImageUrl'] as String?,
-      lifeTime: json['lifeTime'] as bool? ?? false,
-      wantsReminders: json['wantsReminders'] as bool? ?? false,
-    );
-
-Map<String, dynamic> _$WarrantyInfoToJson(WarrantyInfo instance) {
-  final val = <String, dynamic>{
-    'warrantyState': _$WarrantyStateEnumMap[instance.warrantyState],
-    'warrantyId': instance.warrantyId,
-  };
-
-  void writeNotNull(String key, dynamic value) {
-    if (value != null) {
-      val[key] = value;
-    }
-  }
-
-  writeNotNull('name', instance.name);
-  writeNotNull('purchaseDate', instance.purchaseDate?.toIso8601String());
-  writeNotNull('warrantyWebsite', instance.warrantyWebsite);
-  writeNotNull('endOfWarranty', instance.endOfWarranty?.toIso8601String());
-  writeNotNull('reminderDate', instance.reminderDate?.toIso8601String());
-  writeNotNull('details', instance.details);
-  writeNotNull('image', instance.image);
-  writeNotNull('imageUrl', instance.imageUrl);
-  writeNotNull('receiptImage', instance.receiptImage);
-  writeNotNull('receiptImageUrl', instance.receiptImageUrl);
-  val['lifeTime'] = instance.lifeTime;
-  val['wantsReminders'] = instance.wantsReminders;
-  return val;
+extension _$WarrantyInfoAutoequal on WarrantyInfo {
+  @Deprecated(r'Use _$props instead')
+  List<Object?> get _autoequalProps => _$props;
+  List<Object?> get _$props => [
+        id,
+        name,
+        purchaseDate,
+        warrantyWebsite,
+        endOfWarranty,
+        reminderDate,
+        details,
+        image,
+        imageUrl,
+        receiptImage,
+        receiptImageUrl,
+        lifeTime,
+        wantsReminders
+      ];
 }
 
-const _$WarrantyStateEnumMap = {
-  WarrantyState.loading: 'loading',
-  WarrantyState.initial: 'initial',
-  WarrantyState.editing: 'editing',
-  WarrantyState.submitted: 'submitted',
-  WarrantyState.loadingImage: 'loadingImage',
-};
+// **************************************************************************
+// CopyWithGenerator
+// **************************************************************************
+
+abstract class _$WarrantyInfoCWProxy {
+  WarrantyInfo id(String id);
+
+  WarrantyInfo name(String? name);
+
+  WarrantyInfo purchaseDate(DateTime? purchaseDate);
+
+  WarrantyInfo warrantyWebsite(String? warrantyWebsite);
+
+  WarrantyInfo endOfWarranty(DateTime? endOfWarranty);
+
+  WarrantyInfo reminderDate(DateTime? reminderDate);
+
+  WarrantyInfo details(String? details);
+
+  WarrantyInfo image(XFile? image);
+
+  WarrantyInfo imageUrl(String? imageUrl);
+
+  WarrantyInfo receiptImage(XFile? receiptImage);
+
+  WarrantyInfo receiptImageUrl(String? receiptImageUrl);
+
+  WarrantyInfo lifeTime(bool lifeTime);
+
+  WarrantyInfo wantsReminders(bool wantsReminders);
+
+  /// This function **does support** nullification of nullable fields. All `null` values passed to `non-nullable` fields will be ignored. You can also use `WarrantyInfo(...).copyWith.fieldName(...)` to override fields one at a time with nullification support.
+  ///
+  /// Usage
+  /// ```dart
+  /// WarrantyInfo(...).copyWith(id: 12, name: "My name")
+  /// ````
+  WarrantyInfo call({
+    String? id,
+    String? name,
+    DateTime? purchaseDate,
+    String? warrantyWebsite,
+    DateTime? endOfWarranty,
+    DateTime? reminderDate,
+    String? details,
+    XFile? image,
+    String? imageUrl,
+    XFile? receiptImage,
+    String? receiptImageUrl,
+    bool? lifeTime,
+    bool? wantsReminders,
+  });
+}
+
+/// Proxy class for `copyWith` functionality. This is a callable class and can be used as follows: `instanceOfWarrantyInfo.copyWith(...)`. Additionally contains functions for specific fields e.g. `instanceOfWarrantyInfo.copyWith.fieldName(...)`
+class _$WarrantyInfoCWProxyImpl implements _$WarrantyInfoCWProxy {
+  const _$WarrantyInfoCWProxyImpl(this._value);
+
+  final WarrantyInfo _value;
+
+  @override
+  WarrantyInfo id(String id) => this(id: id);
+
+  @override
+  WarrantyInfo name(String? name) => this(name: name);
+
+  @override
+  WarrantyInfo purchaseDate(DateTime? purchaseDate) =>
+      this(purchaseDate: purchaseDate);
+
+  @override
+  WarrantyInfo warrantyWebsite(String? warrantyWebsite) =>
+      this(warrantyWebsite: warrantyWebsite);
+
+  @override
+  WarrantyInfo endOfWarranty(DateTime? endOfWarranty) =>
+      this(endOfWarranty: endOfWarranty);
+
+  @override
+  WarrantyInfo reminderDate(DateTime? reminderDate) =>
+      this(reminderDate: reminderDate);
+
+  @override
+  WarrantyInfo details(String? details) => this(details: details);
+
+  @override
+  WarrantyInfo image(XFile? image) => this(image: image);
+
+  @override
+  WarrantyInfo imageUrl(String? imageUrl) => this(imageUrl: imageUrl);
+
+  @override
+  WarrantyInfo receiptImage(XFile? receiptImage) =>
+      this(receiptImage: receiptImage);
+
+  @override
+  WarrantyInfo receiptImageUrl(String? receiptImageUrl) =>
+      this(receiptImageUrl: receiptImageUrl);
+
+  @override
+  WarrantyInfo lifeTime(bool lifeTime) => this(lifeTime: lifeTime);
+
+  @override
+  WarrantyInfo wantsReminders(bool wantsReminders) =>
+      this(wantsReminders: wantsReminders);
+
+  @override
+
+  /// This function **does support** nullification of nullable fields. All `null` values passed to `non-nullable` fields will be ignored. You can also use `WarrantyInfo(...).copyWith.fieldName(...)` to override fields one at a time with nullification support.
+  ///
+  /// Usage
+  /// ```dart
+  /// WarrantyInfo(...).copyWith(id: 12, name: "My name")
+  /// ````
+  WarrantyInfo call({
+    Object? id = const $CopyWithPlaceholder(),
+    Object? name = const $CopyWithPlaceholder(),
+    Object? purchaseDate = const $CopyWithPlaceholder(),
+    Object? warrantyWebsite = const $CopyWithPlaceholder(),
+    Object? endOfWarranty = const $CopyWithPlaceholder(),
+    Object? reminderDate = const $CopyWithPlaceholder(),
+    Object? details = const $CopyWithPlaceholder(),
+    Object? image = const $CopyWithPlaceholder(),
+    Object? imageUrl = const $CopyWithPlaceholder(),
+    Object? receiptImage = const $CopyWithPlaceholder(),
+    Object? receiptImageUrl = const $CopyWithPlaceholder(),
+    Object? lifeTime = const $CopyWithPlaceholder(),
+    Object? wantsReminders = const $CopyWithPlaceholder(),
+  }) {
+    return WarrantyInfo(
+      id: id == const $CopyWithPlaceholder() || id == null
+          ? _value.id
+          // ignore: cast_nullable_to_non_nullable
+          : id as String,
+      name: name == const $CopyWithPlaceholder()
+          ? _value.name
+          // ignore: cast_nullable_to_non_nullable
+          : name as String?,
+      purchaseDate: purchaseDate == const $CopyWithPlaceholder()
+          ? _value.purchaseDate
+          // ignore: cast_nullable_to_non_nullable
+          : purchaseDate as DateTime?,
+      warrantyWebsite: warrantyWebsite == const $CopyWithPlaceholder()
+          ? _value.warrantyWebsite
+          // ignore: cast_nullable_to_non_nullable
+          : warrantyWebsite as String?,
+      endOfWarranty: endOfWarranty == const $CopyWithPlaceholder()
+          ? _value.endOfWarranty
+          // ignore: cast_nullable_to_non_nullable
+          : endOfWarranty as DateTime?,
+      reminderDate: reminderDate == const $CopyWithPlaceholder()
+          ? _value.reminderDate
+          // ignore: cast_nullable_to_non_nullable
+          : reminderDate as DateTime?,
+      details: details == const $CopyWithPlaceholder()
+          ? _value.details
+          // ignore: cast_nullable_to_non_nullable
+          : details as String?,
+      image: image == const $CopyWithPlaceholder()
+          ? _value.image
+          // ignore: cast_nullable_to_non_nullable
+          : image as XFile?,
+      imageUrl: imageUrl == const $CopyWithPlaceholder()
+          ? _value.imageUrl
+          // ignore: cast_nullable_to_non_nullable
+          : imageUrl as String?,
+      receiptImage: receiptImage == const $CopyWithPlaceholder()
+          ? _value.receiptImage
+          // ignore: cast_nullable_to_non_nullable
+          : receiptImage as XFile?,
+      receiptImageUrl: receiptImageUrl == const $CopyWithPlaceholder()
+          ? _value.receiptImageUrl
+          // ignore: cast_nullable_to_non_nullable
+          : receiptImageUrl as String?,
+      lifeTime: lifeTime == const $CopyWithPlaceholder() || lifeTime == null
+          ? _value.lifeTime
+          // ignore: cast_nullable_to_non_nullable
+          : lifeTime as bool,
+      wantsReminders: wantsReminders == const $CopyWithPlaceholder() ||
+              wantsReminders == null
+          ? _value.wantsReminders
+          // ignore: cast_nullable_to_non_nullable
+          : wantsReminders as bool,
+    );
+  }
+}
+
+extension $WarrantyInfoCopyWith on WarrantyInfo {
+  /// Returns a callable class that can be used as follows: `instanceOfWarrantyInfo.copyWith(...)` or like so:`instanceOfWarrantyInfo.copyWith.fieldName(...)`.
+  // ignore: library_private_types_in_public_api
+  _$WarrantyInfoCWProxy get copyWith => _$WarrantyInfoCWProxyImpl(this);
+}
