@@ -14,9 +14,34 @@ class DetailsImageCard extends StatelessWidget {
       decoration: BoxDecoration(
         borderRadius: BorderRadius.circular(25),
       ),
-      child: Image.network(
-        url,
-        height: MediaQuery.of(context).size.height * .3,
+      child: AnimatedSize(
+        duration: const Duration(milliseconds: 150),
+        child: Container(
+          decoration: BoxDecoration(
+            borderRadius: BorderRadius.circular(25),
+          ),
+          child: Image.network(
+            url,
+            fit: BoxFit.fill,
+            height: MediaQuery.of(context).size.height * .3,
+            loadingBuilder: (BuildContext context, Widget child,
+                ImageChunkEvent? loadingProgress) {
+              if (loadingProgress == null) {
+                return child;
+              }
+              return Container(
+                alignment: Alignment.center,
+                height: MediaQuery.of(context).size.height * .3,
+                child: CircularProgressIndicator(
+                  value: loadingProgress.expectedTotalBytes != null
+                      ? loadingProgress.cumulativeBytesLoaded /
+                          loadingProgress.expectedTotalBytes!
+                      : null,
+                ),
+              );
+            },
+          ),
+        ),
       ),
     );
   }
